Add files via upload
 I have added comments, headers, and the following methods to the abstract class:
-add(value)
-add(index, value)
-addAll(List)
-clear()
-checkIndex()

Please note that I have changed the access modifier of checkIndex() to protected as it is currently used in the subclasses. It is also important to note that the ListClient3 will require a second list to be passed into the process list method in order to run the addAll method.